feat: refactor oll validation logic to separate file
